refactor(rust): extract window sizing into window.rs (slice 1/9)
Move target_window_size + its test out of lib.rs into a dedicated
window module (test now lives beside the code). lib.rs re-exports it
so run()/setup is unchanged. First slice of the backend module split.
